added skeletal animation functionality and new exporter for an improved model format (animation export is broken)

+++ Engine/Graphics/sgObjectFiles.h
@@ -63,7 +63,7 @@ namespace sgObjectFiles
 		unsigned char vtxsize;
 		
 		unsigned short vertexnum;
-		unsigned short indexnum;
+		unsigned int indexnum;
 		float *vertices;
 		unsigned short *indices;
 		unsigned short *bonemapping;
